Add new privileges for reading and writing prinicpals.
This is a work in progress.
