+++ src/group/controller/new.ts
@@ -1,6 +1,6 @@
 import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
-import * as principalService from '../../principal/service';
+import { PrincipalService } from '../../principal/privileged-service';
 import { createGroupForm } from '../formats/html';
 import { uuidUrn } from '../../crypto';
 
@@ -23,6 +23,7 @@ class CreateGroupController extends Controller {
 
   async post(ctx: Context) {
 
+    const principalService = new PrincipalService(ctx.privileges);
     ctx.request.validate<GroupNewForm>('https://curveballjs.org/schemas/a12nserver/group-new-form.json');
 
     ctx.privileges.require('admin');
