+++ src/oauth2/controller/user-access-token.ts
@@ -1,24 +1,18 @@
 import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
-import { Forbidden, BadRequest } from '@curveball/http-errors';
 
 import log from '../../log/service';
 import { EventType } from '../../log/types';
-import * as principalService from '../../principal/service';
+import { PrincipalService } from '../../principal/privileged-service';
 import * as oauth2Service from '../service';
 
 class UserAccessTokenController extends Controller {
 
   async post(ctx: Context<any>) {
 
-    const user = await principalService.findByExternalId(ctx.params.id);
-    if (ctx.auth.equals(user) && !ctx.privileges.has('admin')) {
-      throw new Forbidden('You can only generate OAuth2 access tokens for yourself with this endpoint (unless you have the \'admin\' privilege (which you haven\'t))');
-    }
+    const principalService = new PrincipalService(ctx.privileges);
+    const user = await principalService.findByExternalId(ctx.params.id, 'user');
 
-    if (user.type !== 'user') {
-      throw new BadRequest('This API can only be used for principals of type \'user\'');
-    }
     const token = await oauth2Service.generateTokenDeveloperToken({
       principal: user,
     });
