+++ src/one-time-token/controller/generate.ts
@@ -1,6 +1,6 @@
 import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
-import * as principalService from '../../principal/service';
+import { PrincipalService } from '../../principal/privileged-service';
 import { createToken } from '../service';
 import * as hal from '../formats/hal';
 import { resolve } from 'url';
@@ -11,6 +11,7 @@ class OneTimeTokenController extends Controller {
 
     ctx.privileges.require('admin');
 
+    const principalService = new PrincipalService(ctx.privileges);
     const user = await principalService.findByExternalId(ctx.params.id, 'user');
 
     const token = await createToken(user);
