+++ src/user-app-permissions/controller/user-collection.ts
@@ -1,18 +1,15 @@
 import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
 import * as hal from '../formats/hal';
-import { Forbidden } from '@curveball/http-errors';
-import * as principalService from '../../principal/service';
 import * as userAppPermissionService from '../service';
+import { PrincipalService } from '../../principal/privileged-service';
 
 class UserAppPermissionsCollection extends Controller {
 
   async get(ctx: Context) {
 
+    const principalService = new PrincipalService(ctx.privileges);
     const principal = await principalService.findByExternalId(ctx.params.id, 'user');
-    if (ctx.auth.equals(principal) && !ctx.privileges.has('admin')) {
-      throw new Forbidden('You can only use this API for yourself, or if you have \'admin\' privileges');
-    }
 
     const appPermissions = await userAppPermissionService.findByUser(principal);
     ctx.response.body = hal.collection(
