+++ src/user-app-permissions/controller/user-item.ts
@@ -2,14 +2,16 @@ import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
 import * as hal from '../formats/hal';
 import { Forbidden } from '@curveball/http-errors';
-import * as principalService from '../../principal/service';
+import { PrincipalService } from '../../principal/privileged-service';
 import * as userAppPermissionService from '../service';
 
 class UserAppPermissionsItem extends Controller {
 
   async get(ctx: Context) {
 
+    const principalService = new PrincipalService(ctx.privileges);
     const user = await principalService.findByExternalId(ctx.params.id, 'user');
+
     if (ctx.auth.equals(user) && !ctx.privileges.has('admin')) {
       throw new Forbidden('You can only use this API for yourself, or if you have \'admin\' privileges');
     }
