+++ src/user/controller/by-href.ts
@@ -1,13 +1,13 @@
 import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
-import * as principalService from '../../principal/service';
+import { PrincipalService } from '../../principal/privileged-service';
 
 class UserByHrefController extends Controller {
 
   async get(ctx: Context) {
 
+    const principalService = new PrincipalService(ctx.privileges);
     const principal = await principalService.findByHref(decodeURIComponent(ctx.params.href));
-
     ctx.redirect(303, principal.href);
 
   }
