+++ src/user/controller/collection.ts
@@ -1,8 +1,8 @@
 import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
-import { BadRequest, Conflict, Forbidden, NotFound } from '@curveball/http-errors';
+import { BadRequest, Conflict, NotFound } from '@curveball/http-errors';
 import * as hal from '../formats/hal';
-import * as principalService from '../../principal/service';
+import { PrincipalService } from '../../principal/privileged-service';
 
 type NewPrincipalBody = {
   nickname: string;
@@ -14,16 +14,13 @@ class UserCollectionController extends Controller {
 
   async get(ctx: Context) {
 
+    const principalService = new PrincipalService(ctx.privileges);
     const users = await principalService.findAll('user');
     ctx.response.body = hal.collection(users);
 
   }
 
   async post(ctx: Context) {
-
-    if (!ctx.privileges.has('admin')) {
-      throw new Forbidden('Only users with the "admin" privilege may create new users');
-    }
 
     ctx.request.validate<NewPrincipalBody>(
       'https://curveballjs.org/schemas/a12nserver/principal-new.json'
@@ -33,6 +30,8 @@ class UserCollectionController extends Controller {
     if (!identity) {
       throw new BadRequest('You must specify a link with rel "me", either via a HAL link or HTTP Link header');
     }
+
+    const principalService = new PrincipalService(ctx.privileges);
 
     try {
       await principalService.findByIdentity(identity);
