+++ src/user/controller/item.ts
@@ -5,9 +5,8 @@ import * as userHal from '../formats/hal';
 import * as appHal from '../../app/formats/hal';
 import * as groupHal from '../../group/formats/hal';
 import * as userService from '../service';
-import * as principalService from '../../principal/service';
 import * as groupService from '../../group/service';
-import { Forbidden } from '@curveball/http-errors';
+import { PrincipalService } from '../../principal/privileged-service';
 
 type EditPrincipalBody = {
   nickname: string;
@@ -29,6 +28,7 @@ class UserController extends Controller {
 
   async get(ctx: Context) {
 
+    const principalService = new PrincipalService(ctx.privileges);
     const principal = await principalService.findByExternalId(ctx.params.id);
 
     let hasControl = false;
@@ -84,12 +84,10 @@ class UserController extends Controller {
 
   async put(ctx: Context) {
 
-    if (!ctx.privileges.has('admin')) {
-      throw new Forbidden('Only users with the "admin" privilege may edit users');
-    }
     ctx.request.validate<EditPrincipalBody>(
       'https://curveballjs.org/schemas/a12nserver/principal-edit.json'
     );
+    const principalService = new PrincipalService(ctx.privileges);
 
     const user = await principalService.findByExternalId(ctx.params.id);
     user.active = !!ctx.request.body.active;
